+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- `CueEnterEvent` and `CueExitEvent` to signal when a subtitle entry transitions into an active or inactive status respectively
+
+## [0.18.0] (2024-03-06)
+
+### Changed
+
+- React Native version to `0.73.4`
+- Update Bitmovin's native Android SDK version to `3.61.0`
+- Update Bitmovin's native iOS SDK version to `3.56.3`
+
+### Fixed
+
+- Android: Subtitles appear out of frame
+
 ## [0.17.0] (2024-01-26)
 
 ### Added